Rename warehouse.utils.proxy to warehouse.utils.wsgi
